+++ zp-core/zp-extensions/slideshow2.php
@@ -66,9 +66,10 @@ class cycle {
 			setOptionDefault('cycle-slideshow_colorbox_transition', 'fade');
 			setOptionDefault('cycle-slideshow_colorbox_imagetype', 'sizedimage');
 			setOptionDefault('cycle-slideshow_colorbox_imagetitle', 1);
-
-			cacheManager::deleteThemeCacheSizes('cycle');
-			cacheManager::addThemeCacheSize('cycle', NULL, getOption('cycle-slideshow_width'), getOption('cycle-slideshow_height'), NULL, NULL, NULL, NULL, NULL, NULL, NULL, true);
+			if (class_exists('cacheManager')) {
+				cacheManager::deleteThemeCacheSizes('cycle');
+				cacheManager::addThemeCacheSize('cycle', NULL, getOption('cycle-slideshow_width'), getOption('cycle-slideshow_height'), NULL, NULL, NULL, NULL, NULL, NULL, NULL, true);
+			}
 		}
 	}
